+++ app/src/main/java/com/example/weathercompose/model/Condition.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Condition(
     @SerializedName("code")
-    val code: Int,
+    val code: Double,
     @SerializedName("icon")
     val icon: String,
     @SerializedName("text")
